feat: :alembic: try freebsd

+++ osbased/time_linux.go
@@ -1,4 +1,4 @@
-//go:build (amd64 || arm64) && linux
+//go:build linux
 
 package osbased
 
@@ -14,12 +14,12 @@ func ModTime(a os.FileInfo) time.Time {
 
 func AccessTime(a os.FileInfo) time.Time {
 	atim := a.Sys().(*syscall.Stat_t).Atim
-	return time.Unix(atim.Sec, atim.Nsec)
+	return time.Unix(int64(atim.Sec), int64(atim.Nsec))
 }
 
 func CreateTime(a os.FileInfo) time.Time {
 	ctim := a.Sys().(*syscall.Stat_t).Ctim
-	return time.Unix(ctim.Sec, ctim.Nsec)
+	return time.Unix(int64(ctim.Sec), int64(ctim.Nsec))
 }
 
 func BirthTime(a os.FileInfo) time.Time {
